don't send key event when the textinput has client
A little work-around.
Waiting for https://github.com/go-flutter-desktop/go-flutter/issues/314
to have a real answer.

For now, I am a bit sceptical on how the keyboard text
selection/keyboard pasting will be handled in the flutter side.

I'm afraid it might not be suited to our need. We are sending
GLFW(linux) related event to the framework. But the framework only has
one handler, linux. This means, sending keyevent on darwin will not
support the correct keyboard layout (Ctrl -> Meta).

As those different keyboard layout are already supported in go-flutter,
this work-around ensure that we keep control of the textinput when
needed.